LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/math-font-script-src.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/gen...


FAIL Speculative parsing, page load: math-font-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/generated/page-load/resources/math-font-script-src-framed.sub.html?uuid=b7882d2d-5ab1-497d-b4af-f6140feeb3e2"
